Refactor and parameterize DB variables
